+++ quantities.hs
@@ -11,9 +11,9 @@ parse ["-h"] = usage   >> exit
 parse ["-v"] = version >> exit
 parse []     = usage   >> exit
 parse [s]    = putStr $ (++ "\n") $ either show show $ fromString s
-parse _      = usage   >> die
+parse _      = usage   >> Main.die
 
 usage   = putStrLn "Usage: quantities [-vh] expression"
 version = putStrLn "Haskell quantities 0.3.0"
-exit    = exitWith ExitSuccess
+exit    = exitSuccess
 die     = exitWith (ExitFailure 1)